pyproject.toml
[project]
name = "android-ai-service"
version = "0.1.0"
description = "AI Service for food ordering app - Multi-Agent Chatbot & Vietnamese ASR"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "langchain>=0.3",
    "langgraph>=0.3",
    "langchain-openai>=0.3",
    "langchain-ollama>=0.3",
    "chunkformer>=0.1",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "soundfile>=0.13",
    "sqlalchemy>=2.0",
    "aiomysql>=0.2",
    "silero-vad>=5.1",
]
